backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sentence-transformers
pyjwt
passlib
bcrypt==4.0.1
google-generativeai
numpy
pytesseract
python-multipart
pypdf
langchain
langchain-community
pillow
itsdangerous
authlib
httpx